Add files via upload
This project aims to examine the characteristics of customers through data from marketers who sell products by mail only. Subsequently, the analysis aims to predict the consumption of customers and identify the criteria for selecting customers to send promotional materials in the future from the perspective of a direct mail marketer.